LayoutTests/imported/w3c/web-platform-tests/css/css-images/parsing/gradient-interpolation-method-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Images Module Level 4: parsing gradients with color interpolation methods</title>
<link rel="author" title="Sam Weinig" href="mailto:weinig@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-images-4/#gradients">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#color-interpolation-method">
<meta name="assert" content="gradients supports the addition of color-interpolation-method to the grammar">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>

const LINEAR_GRADIENT_SPECIFIERS = [
    { input: '30deg' },
    { input: '0', output: '0deg' }, // unitless 0 special case
    { input: 'to left' },
    { input: 'to right bottom' },
    { input: 'to bottom right', output: 'to right bottom' },
];

const RADIAL_GRADIENT_SPECIFIERS = [
    { input: '50px' },
    { input: '50px at left top' },
    { input: 'circle 50px', output: '50px' },
    { input: 'circle 50px at left', output: '50px at left center' },
    { input: '50px 50%' },
    { input: 'ellipse 50% 40em', output: '50% 40em' },
    { input: 'at right center' },
];

const CONIC_GRADIENT_SPECIFIERS = [
    { input: 'from 30deg' },
    { input: 'from 30deg at left', output: 'from 30deg at left center' },
    { input: 'at left', output: 'at left center' },
    { input: 'from 30deg at center 10%' },
    { input: 'at center 10%' },
    { input: 'from 30deg at right 20% top 30%' },
    { input: 'at right 20% top 30%' },
    { input: 'from 30deg at left 10px top 50em' },
    { input: 'at left 10px top 50em' },
];


function test_gradients(gradientFunction, colorInterpolationMethod, colorInterpolationMethodResult, specifiers) {
    const resultForNoSpecifierCase = colorInterpolationMethodResult == "" ? "" : `in ${colorInterpolationMethodResult}, `
    test_computed_value(`background-image`, `${gradientFunction}(in ${colorInterpolationMethod}, red, blue)`, `${gradientFunction}(${resultForNoSpecifierCase}red, blue)`)

    for (const specifier of specifiers) {
        const input = specifier.input
        const output = specifier.output ? specifier.output : specifier.input
        const result = colorInterpolationMethodResult == "" ? ", " : ` in ${colorInterpolationMethodResult}, `
        test_computed_value(`background-image`, `${gradientFunction}(${input} in ${colorInterpolationMethod}, red, blue)`, `${gradientFunction}(${output}${result}red, blue)`)
        test_computed_value(`background-image`, `${gradientFunction}(in ${colorInterpolationMethod} ${input}, red, blue)`, `${gradientFunction}(${output}${result}red, blue)`)
    }
}

function test_each_interpolation_method(gradientFunction, specifiers) {
    test_gradients(gradientFunction, "srgb", "", specifiers)

    for (const colorSpace of [ "lab", "oklab", "srgb-linear", "xyz", "xyz-d50", "xyz-d65" ]) {
        const colorInterpolationMethod = colorSpace
        const colorInterpolationMethodResult = colorSpace == "xyz" ? "xyz-d65" : colorInterpolationMethod

        test_gradients(gradientFunction, colorInterpolationMethod, colorInterpolationMethodResult, specifiers)
    }

    for (const colorSpace of [ "hsl", "hwb", "lch", "oklch" ]) {
        for (const hueInterpolationMethod of [ "", " shorter hue", " longer hue", " increasing hue", " decreasing hue", " specified hue" ]) {
            const colorInterpolationMethod = `${colorSpace}${hueInterpolationMethod}`
            const colorInterpolationMethodResult = hueInterpolationMethod == " shorter hue" ? colorSpace : colorInterpolationMethod

            test_gradients(gradientFunction, colorInterpolationMethod, colorInterpolationMethodResult, specifiers)
        }
    }
}

test_each_interpolation_method("linear-gradient", LINEAR_GRADIENT_SPECIFIERS)
test_each_interpolation_method("radial-gradient", RADIAL_GRADIENT_SPECIFIERS)
test_each_interpolation_method("conic-gradient", CONIC_GRADIENT_SPECIFIERS)

</script>
</body>
</html>
